Allow reusing DSP
This patch only adds namespaces around the DSP classes, so all the 4
variants can be pulled from other projects without causing ODR
violations.

This modification renders the original compilation useless, it just
makes possible to compile the DSP sources separately on its own
C++ namespace.

The only reason to fork instead of maintaining a patch on the relevant
codebase, is to avoid the chore that is to make  the "patch" binary to
be available on Windows builds.